handle context deadline error
