+++ src/internal/connector/graph/errors.go
@@ -2,6 +2,7 @@ package graph
 
 import (
 	"context"
+	"net/http"
 	"net/url"
 	"os"
 
@@ -95,7 +96,7 @@ func IsErrTimeout(err error) bool {
 		return true
 	}
 
-	if errors.Is(err, context.DeadlineExceeded) || os.IsTimeout(err) {
+	if errors.Is(err, context.DeadlineExceeded) || os.IsTimeout(err) || errors.Is(err, http.ErrHandlerTimeout) {
 		return true
 	}
